feature 브랜치 mypage 생성

+++ src/main/java/com/soldesk6F/ondal/user/service/UserService.java
@@ -2,7 +2,6 @@ package com.soldesk6F.ondal.user.service;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.URLDecoder;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -13,10 +12,10 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.ui.Model;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.soldesk6F.ondal.login.CustomUserDetails;
 import com.soldesk6F.ondal.user.entity.User;
 import com.soldesk6F.ondal.user.repository.UserRepository;
 
@@ -29,7 +28,6 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class UserService {
 
-	
     private final UserRepository userRepository;
     private final BCryptPasswordEncoder passwordEncoder;
 
@@ -50,6 +48,10 @@ public class UserService {
     
     public boolean isPhoneDuplicate(String userPhone) {
         return userRepository.existsByUserPhone(userPhone);
+    }
+    
+    public boolean isNicknameDuplicate(String nickName) {
+    	return userRepository.existsByNickName(nickName);
     }
 
     public boolean registerUser(String userId, String userName, String nickname, String email,
@@ -87,7 +89,8 @@ public class UserService {
 	                .password(encryptedPassword)
 	                .userPhone(userPhone)
 	                .userSelectedAddress(userSelectedAddress)	
-	                .userProfile(filePath)
+//	                .userProfile(filePath)
+	                .userProfile(fileName)
 	                .socialLoginProvider(socialLoginProvider)
 	                .build();
 	
@@ -100,74 +103,56 @@ public class UserService {
 	}
     
     @Transactional
-    public boolean updatePassword(String oldPassword, String password, User user, RedirectAttributes rAttr) {
-    	Optional<User> findUser = userRepository.findById(user.getUserUuid());
-    	try {
-    		User currentUser = findUser.get();
-    		if (passwordEncoder.matches(oldPassword, currentUser.getPassword())) {
-    			if (password.equals(oldPassword)) {
-    				rAttr.addFlashAttribute("result", 1);
-    				rAttr.addFlashAttribute("resultMsg", "비밀번호 변경사항이 없습니다." );
-    				return false;
-    			}
-    			String encryptedPassword = passwordEncoder.encode(password);
-    			
-    			currentUser.setPassword(encryptedPassword);
-//    			rAttr.addFlashAttribute("result", 0);
-//    			rAttr.addFlashAttribute("resultMsg", "비밀번호 일치");
-    			return true;
-    		}
-    		rAttr.addFlashAttribute("result", 1);
-    		rAttr.addFlashAttribute("resultMsg", "기존 비밀번호가 일치하지 않습니다.");
-    		return false;
-		} catch (Exception e) {
-			e.printStackTrace();
-			return false;
-		}
-    }
-    
-    
-    @Transactional
-    public boolean updateUserNickname(String nickName, User user, Model model) {
-    	Optional<User> findUser = userRepository.findByUserId(user.getUserId());
+    public boolean updateUserNickname(CustomUserDetails cud, String nickName, RedirectAttributes rAttr) {
+    	Optional<User> findUser = userRepository.findByUserId(cud.getUser().getUserId());
     	
     	if (findUser.isEmpty()) {
-    		throw new IllegalArgumentException("존재하지 않는 아이디 입니다.");
+    		rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "존재하지 않는 ID입니다.");
+    		return false;
     	}
-    	
-    	User currentUser = findUser.get();
-    	
-   		if (currentUser.getNickName().equals(nickName)) {
-    		model.addAttribute("error", "기존 닉네임과 동일합니다.");
+   		if (findUser.get().getNickName().equals(nickName)) {
+   			rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "닉네임을 변경할 수 없습니다.");
     		return false;
     	} else {
-//    		findUser.ifPresent(U -> U.setNickName(nickName) );
-//    		findUser.ifPresent(U -> U.setUpdatedDate(LocalDateTime.now()));
-    		currentUser.setNickName(nickName);
-    		currentUser.setUpdatedDate(LocalDateTime.now());
+    		findUser.get().updateNickname(nickName);
+//    		findUser.get().setUpdatedDate(LocalDateTime.now());
+    		findUser.get().updateUpdatedDate(LocalDateTime.now());
+    		cud.getUser().setNickName(nickName);
+    		rAttr.addFlashAttribute("result", 0);
+    	    rAttr.addFlashAttribute("resultMsg", "닉네임 변경 성공!");
     		return true;
     	}
     }
     
-    
     @Transactional
-    public boolean updateUserPicture(User user, MultipartFile profileImage, Model model) {
-    	Optional<User> findUser = userRepository.findByUserId(user.getUserId());
+    public boolean updateUserPicture(CustomUserDetails cud, 
+    		MultipartFile profileImage, 
+    		RedirectAttributes rAttr) {
+    	Optional<User> findUser = userRepository.findByUserId(cud.getUser().getUserId());
     		
+    	if (findUser.isEmpty()) {
+    		rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "존재하지 않는 ID입니다.");
+    		return false;
+    	}
+    	
    		try {
    			String old_profImgName = findUser.get().getUserProfile();
 //    		String old_profImgName = old_profImgPath.split("\\\\")[1];
    			if (!old_profImgName.equals("default.png")) {
    				String oldSavePath = new File(uploadDir).getAbsolutePath();
    				Path oldImgPath = Paths.get(oldSavePath, old_profImgName);
-   				//System.out.println(oldImgPath.toString());
    				if (Files.exists(oldImgPath)) {
    					Files.delete(oldImgPath);
    				}
    			}
 		} catch (Exception e) {
 			e.printStackTrace();
-			model.addAttribute("error", "존재하지 않는 파일입니다.");
+			rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "파일이 존재하지 않습니다.");
+			return false;
 		}
     	
     	try {
@@ -179,7 +164,6 @@ public class UserService {
 	        	String originalFilename = profileImage.getOriginalFilename();
 	        	extension = originalFilename.substring(originalFilename.lastIndexOf(".") + 1);
 	        	fileName = findUser.get().getUserId() + "_" + System.currentTimeMillis() + "." + extension;
-//	        	fileName = findUser.get().getUserUuidAsString() + "." + extension;
 	            String savePath = new File(uploadDir).getAbsolutePath(); // getRealPath()
 	            System.out.println(savePath);
 	            File saveFolder = new File(savePath);
@@ -194,45 +178,86 @@ public class UserService {
 	            
 	            //final String finalFilePath = new String(filePath);
 	            final String finalFileName = new String(fileName);
-	            //findUser.ifPresent(U -> U.setUserProfile(finalFilePath));
-	            findUser.ifPresent(U -> U.setUserProfile(finalFileName));
-	            findUser.ifPresent(U -> U.setUpdatedDate(LocalDateTime.now()));
+//	            model.addAttribute("fileName", finalFileName);
+	            findUser.get().updateProfile(finalFileName);
+	            findUser.get().updateUpdatedDate(LocalDateTime.now());
+	            cud.getUser().setUserProfile(finalFileName);
+	            rAttr.addFlashAttribute("result", 0);
+				rAttr.addFlashAttribute("resultMsg", "프로필 이미지 변경 성공!");
 	            return true;
 	        }
 	        final String finalFileName = new String(fileName);
-	        findUser.ifPresent(U -> U.setUserProfile(finalFileName));
-	        findUser.ifPresent(U -> U.setUpdatedDate(LocalDateTime.now()));
+	        findUser.get().updateProfile(finalFileName);
+	        findUser.get().updateUpdatedDate(LocalDateTime.now());
+	        cud.getUser().setUserProfile(finalFileName);
+			rAttr.addFlashAttribute("result", 0);
+			rAttr.addFlashAttribute("resultMsg", "기본 이미지로 변경 성공!");
 	        return true;
 		} catch (Exception e) {
 			e.printStackTrace();
+			rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "프로필 이미지 변경 실패!");
 			return false;
 		}
     	
     }
     
     @Transactional
-    public boolean updateUserPhone(String userPhone, User user, Model model) {
-    	Optional<User> findUser = userRepository.findByUserId(user.getUserId());
-    	
+    public boolean updateUserPhone(CustomUserDetails cud, String userPhone, RedirectAttributes rAttr) {
+    	Optional<User> findUser = userRepository.findByUserId(cud.getUser().getUserId());
+
     	if (findUser.isEmpty()) {
-    		throw new IllegalArgumentException("존재하지 않는 아이디 입니다.");
+    		rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "존재하지 않는 ID입니다.");
+			return false;
     	}
-    	
-    	User currentUser = findUser.get();
-    	
-   		if (currentUser.getUserPhone().equals(userPhone)) {
-    		model.addAttribute("error", "기존 전화번호와 동일합니다.");
+   		if (findUser.get().getUserPhone().equals(userPhone)) {
+			rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "전화번호 변경 사항이 없습니다.");
     		return false;
     	} else {
-//    		findUser.ifPresent(U -> U.setNickName(nickName) );
-//    		findUser.ifPresent(U -> U.setUpdatedDate(LocalDateTime.now()));
-    		currentUser.setUserPhone(userPhone);
-    		currentUser.setUpdatedDate(LocalDateTime.now());
+    		findUser.get().updatePhone(userPhone);
+    		findUser.get().updateUpdatedDate(LocalDateTime.now());
+    		cud.getUser().setUserPhone(userPhone);
+    		rAttr.addFlashAttribute("result", 0);
+    		rAttr.addFlashAttribute("resultMsg", "전화번호 변경 성공!");
     		return true;
     	}
     }
     
-    
+    @Transactional
+    public boolean updatePassword(CustomUserDetails cud, String oldPassword, String password, RedirectAttributes rAttr) {
+    	Optional<User> findUser = userRepository.findByUserId(cud.getUser().getUserId());
+    	
+    	if (findUser.isEmpty()) {
+    		rAttr.addFlashAttribute("result", 1);
+			rAttr.addFlashAttribute("resultMsg", "존재하지 않는 ID입니다.");
+    		return false;
+    	}
+    	
+    	try {
+   			if (passwordEncoder.matches(oldPassword, findUser.get().getPassword())) {
+    			if (password.equals(oldPassword)) {
+    				rAttr.addFlashAttribute("result", 1);
+    				rAttr.addFlashAttribute("resultMsg", "비밀번호 변경사항이 없습니다." );
+    				return false;
+    			}
+    			String encryptedPassword = passwordEncoder.encode(password);
+    			findUser.get().updatePassword(encryptedPassword);
+    			findUser.get().updateUpdatedDate(LocalDateTime.now());
+    			cud.getUser().setPassword(password);
+    			rAttr.addFlashAttribute("result", 0);
+    			rAttr.addFlashAttribute("resultMsg", "비밀번호 변경 성공!");
+    			return true;
+    		}
+    		rAttr.addFlashAttribute("result", 1);
+    		rAttr.addFlashAttribute("resultMsg", "기존 비밀번호가 일치하지 않습니다.");
+    		return false;
+		} catch (Exception e) {
+			e.printStackTrace();
+			return false;
+		}
+    }
     
 }
